Handle topology messaging
When a server connects to a peer, it sends an I-AM message giving its
unique ID.

When a server receives an I-AM message, it updates its peer information.

When a server's peer-set changes, it'll broadcast its new connectivity
across the mesh network as an I-SEE message. As these are received,
servers update their own copy of the map.

They throw away map data about servers that are no longer reachable;
thus the network should rapidly converge on a new, accurate topology.